# infrastructure/src/main/resources/static/openapi/api-notificaciones.yaml
openapi: 3.1.0
info:
  title: API de Notificaciones
  description: >
    API de Notification-uco. Cada endpoint corresponde a un caso de uso documentado.  Convención de nombramiento: los endpoints
    que representan una acción de negocio explícita (no CRUD genérico) usan el sufijo `:accion` en
    el path (patrón "custom methods", ej. Google AIP-136) en vez de forzar un verbo HTTP que no
    encaja; los que sí son CRUD natural (crear, consultar) usan REST estándar. CU-03 (despachar) no
    tiene endpoint propio -- es un proceso interno disparado por el consumidor de RabbitMQ, nunca
    invocado directamente por un cliente.
  version: 1.0.0
servers:
  - url: /
    description: >
      Relativo al host donde se sirve este contrato -- funciona igual en local, dev y prod sin
      editar este archivo cuando cambie el puerto o el dominio.
...
paths:
  /notifications:
    post:
      tags: [Notificaciones]
      summary: Enviar una notificación
      description: >
        CU-01. Acepta la notificación, la registra en estado PENDING y la encola para su despacho
        asíncrono. Si ya existe una notificación con el mismo (tenantId, externalId), devuelve esa
        misma notificación en vez de crear una segunda (idempotencia).
      operationId: sendNotification
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNotificationRequest'
      responses:
        '202':
          description: Notificación aceptada (nueva o duplicada -- ver el campo "duplicate").
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendNotificationResponse'
        '400':
          description: >
            El canal no existe/está deshabilitado, el contenido no cumple el esquema del canal, o un adjunto
            es inválido: el canal no acepta adjuntos; el adjunto incumple un tope del servicio (cantidad,
            tamaño, tamaño total, tipo, extensión) o una regla del canal; lleva content y url a la vez o
            ninguno; content no es Base64 válido o no coincide con sizeBytes; url no es una subida de este
            tenant o sus datos no coinciden; la subida está INFECTED; el tipo real no coincide con el
            declarado; o el archivo contiene software malicioso. El mensaje identifica el adjunto por su
            posición (attachments[i]) y la regla incumplida; nunca incluye el contenido ni la dirección.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: >
            Un adjunto referencia una subida todavía en PENDING_SCAN. Reintentar cuando
            GET /attachment-uploads/{uploadId} devuelva CLEAN.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: El cuerpo de la solicitud supera 8 MB.
        '503':
          description: >
            El antivirus o el almacén de archivos no están disponibles o no respondieron a tiempo. La
            notificación no se aceptó; puede reintentarse. Las notificaciones sin adjuntos no se ven
            afectadas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Notificaciones]
      summary: Consultar histórico y trazabilidad
      description: >
        CU-05. Busca notificaciones por filtros combinables. Cada resultado incluye su recorrido
        completo de intentos, no solo el estado actual (a diferencia de GET /notifications/{id}).
      operationId: searchNotifications
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: recipientId
          in: query
          schema:
            type: string
        - name: channelType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: from
          in: query
          description: Fecha/hora mínima de aceptación (inclusive).
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          description: Fecha/hora máxima de aceptación (inclusive).
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          description: Tamaño de página. Default 50, máximo 200.
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          description: Posición inicial dentro del conjunto de resultados. Default 0.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Página de notificaciones que cumplen el criterio (items vacío si no hay resultados).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationSearchResponse'
        '400':
          description: >
            Parámetros inválidos -- from posterior a to, o limit/offset fuera de su rango válido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /notifications/{id}:
    get:
      tags: [Notificaciones]
      summary: Consultar el estado de una notificación
      description: CU-04. Devuelve el estado actual, canal, proveedor y marca de tiempo del último intento.
      operationId: getNotificationStatus
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/NotificationId'
      responses:
        '200':
          description: Estado actual de la notificación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationStatusResponse'
        '404':
          description: >
            No existe una notificación con ese id, o no pertenece al tenant del solicitante --
            ambos casos responden igual, para no revelar si el id existe bajo otro cliente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      description: >
        Identificador del tenant solicitante. Placeholder mientras CU-10 (autenticación) sigue
        bloqueado por DEP-01 -- se reemplazará por la identidad resuelta desde el token una vez
        se cierre esa dependencia con el equipo de Seguridad.
      schema:
        type: string
    NotificationId:
      name: id
      in: path
      required: true
      description: Identificador de la notificación.
      schema:
        type: string
        format: uuid
...
  schemas:
    SendNotificationRequest:
      type: object
      required:
        - externalId
        - channelType
        - recipientId
        - recipientAddress
        - body
        - priority
      properties:
        externalId:
          type: string
          description: Identificador propio del sistema cliente, usado para idempotencia.
          example: order-42
        channelType:
          type: string
          description: >
            Canal registrado en el catálogo. La configuración por defecto declara EMAIL, SMS y PUSH.
          example: EMAIL
        recipientId:
          type: string
          description: Identificador del destinatario, estable entre canales.
          example: recipient-1
        recipientAddress:
          type: string
          description: >
            Dirección concreta del canal usado en este envío (email, teléfono, etc.). En SMS es un
            número en formato internacional (signo +, código de país y número, ej. +573001234567); un
            número sin ese formato termina en FAILED sin enviarse. En PUSH es el identificador de
            dispositivo que el proveedor de push emitió a la aplicación cliente (ej. device-token-demo-1234);
            se trata como texto opaco y no se valida su forma: un identificador inválido o caducado termina
            en FAILED sin reintentos.
          example: alice@example.com
        subject:
          type: string
          nullable: true
          description: >
            Opcional -- algunos canales (ej. SMS) no lo usan. En SMS se admite y se ignora: no se
            envía ni se antepone al cuerpo. En PUSH es el título visible, opcional, de hasta 100
            caracteres.
          example: Bienvenida
        body:
          type: string
          description: >
            Cuerpo de la notificación. La longitud máxima la fija la forma de contenido del canal en
            el catálogo; en SMS son 160 caracteres y en PUSH 900 por defecto. Un contenido más largo se
            rechaza con 400 y nunca se trunca.
          example: Hola Alice, tu pedido fue confirmado.
        priority:
          $ref: '#/components/schemas/Priority'
        attachments:
          type: array
          nullable: true
          maxItems: 5
          description: >
            Opcional. Archivos que acompañan la notificación, en el orden en que se entregarán. El canal
            debe declararlos en su forma de contenido (ver GET /channels, campo contentSchema, propiedad
            attachments); un canal que no los declara rechaza toda notificación con adjuntos. Con la
            configuración por defecto ningún canal los declara. Cada adjunto lleva exactamente uno de
            content (archivo de hasta 1 MB codificado en Base64) o url (archivo de más de 1 MB subido antes
            con POST /attachment-uploads y en estado CLEAN). A lo sumo 5 adjuntos y 25 MB (26214400 bytes)
            en total. El servicio verifica el tipo real de cada archivo y lo analiza con un antivirus. Una
            notificación con un adjunto inválido se rechaza completa; nunca se acepta sin el adjunto. Si el
            proveedor que la despacha no sabe enviar adjuntos, termina en FAILED sin enviarse.
          items:
            $ref: '#/components/schemas/Attachment'
    Attachment:
      type: object
      required: [fileName, contentType, sizeBytes]
      oneOf:
        - required: [content]
          not:
            required: [url]
        - required: [url]
          not:
            required: [content]
      properties:
        fileName:
          type: string
          minLength: 1
          maxLength: 255
          description: >
            Nombre con el que lo verá el destinatario. Sin separadores de ruta ni caracteres de control;
            distinto de "." y "..". No puede terminar en una extensión prohibida (sin distinguir
            mayúsculas, ignorando puntos y espacios finales): .exe .msi .bat .cmd .com .scr .pif .vbs .vbe
            .js .jse .wsf .wsh .ps1 .hta .cpl .msc .reg .lnk .jar .dll .sh .apk .app .gadget .com.pif .msix
            .war.
          example: factura-0042.pdf
        contentType:
          type: string
          description: >
            Tipo de medio. Se compara en minúsculas y sin parámetros, y debe coincidir con el tipo real
            del contenido. Tipos admitidos por el servicio: application/pdf, image/png, image/jpeg,
            text/plain, text/csv,
            application/vnd.openxmlformats-officedocument.wordprocessingml.document,
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet. Cada canal puede admitir
            menos.
          example: application/pdf
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
          maximum: 10485760
          description: >
            Tamaño real en bytes. Hasta 1048576 va en content; de 1048577 a 10485760, en url. Con content
            debe ser igual al tamaño decodificado; con url, igual al de la subida. Cada canal puede fijar
            menos. Límite inclusivo.
          example: 183422
        content:
          type: string
          format: byte
          description: >
            Archivo de hasta 1 MB codificado en Base64 estándar (RFC 4648, con relleno), sin prefijo
            "data:" ni saltos de línea. Dato sensible: nunca aparece en respuestas, registros ni eventos.
        url:
          type: string
          format: uri
          maxLength: 2048
          description: >
            uploadUrl devuelta por POST /attachment-uploads para el mismo tenant, cuya subida está en
            estado CLEAN. Cualquier otra dirección se rechaza; el servicio no la descarga. La parte de
            consulta se ignora y la dirección no se guarda.
...
    SendNotificationResponse:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/NotificationStatus'
        duplicate:
          type: boolean
          description: true si ya existía una notificación con el mismo (tenantId, externalId).
...
    NotificationStatusResponse:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/NotificationStatus'
        channelType:
          type: string
        providerId:
          type: string
          nullable: true
          description: Proveedor del último intento de envío. Nulo si aún no se ha intentado despachar.
        lastUpdatedAt:
          type: string
          format: date-time
    NotificationSearchResponse:
      type: object
      required: [items, limit, offset, hasNext]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/NotificationHistoryItem'
        limit:
          type: integer
          description: Tamaño de página efectivamente aplicado.
        offset:
          type: integer
          description: Offset efectivamente aplicado.
        hasNext:
          type: boolean
          description: true si existen más resultados después de esta página.
    NotificationHistoryItem:
      type: object
      properties:
        notificationId:
          type: string
          format: uuid
        externalId:
          type: string
        recipientId:
          type: string
        channelType:
          type: string
        status:
          $ref: '#/components/schemas/NotificationStatus'
        acceptedAt:
          type: string
          format: date-time
        deliveryAttempts:
          type: array
          items:
            $ref: '#/components/schemas/DeliveryAttempt'
    DeliveryAttempt:
      type: object
      properties:
        occurredOn:
          type: string
          format: date-time
        result:
          type: string
          enum: [ACCEPTED, RECOVERABLE_FAILURE, PERMANENT_FAILURE]
        origin:
          type: string
          enum: [MANUAL, AUTOMATIC]
        providerId:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH]
    NotificationStatus:
      type: string
      enum: [PENDING, IN_PROCESS, DELIVERED, RECOVERABLE, FAILED, DISCARDED]
...